requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
cachecontrol==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-filter==23.2
django-rest-auth==0.9.5
django-sslserver==0.22
djangorestframework==3.14.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
jwt==1.3.1
msgpack==1.0.5
mysqlclient==2.2.0
oauthlib==3.2.2
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
